requirements.txt.txt
streamlit==0.89.0
requests==2.26.0
Pillow==8.4.0
fastapi==0.68.0
uvicorn==0.15.0
numpy==1.21.5
tensorflow==2.7.0
